v2 generated with functioning HMAC
