+++ src/main/java/com/messagemedia/api/Configuration.java
@@ -1,0 +1,19 @@
+package com.messagemedia.api;
+
+
+public class Configuration {
+    //The base Uri for API calls
+    public static String baseUri = "https://api.messagemedia.com";
+
+    //The username to use with basic authentication
+    public static String basicAuthUserName = "TODO: Replace";
+
+    //The password to use with basic authentication
+    public static String basicAuthPassword = "TODO: Replace";
+
+    //The username to use with HMAC authentication
+    public static String hmacAuthUserName = "";
+
+    //The password to use with HMAC authentication
+    public static String hmacAuthPassword = "";
+}
