+++ src/main/java/com/messagemedia/api/controllers/BaseController.java
@@ -1,0 +1,80 @@
+package com.messagemedia.api.controllers;
+
+import com.messagemedia.api.exceptions.*;
+import com.messagemedia.api.http.client.HttpClient;
+import com.messagemedia.api.http.client.HttpContext;
+import com.messagemedia.api.http.client.HttpCallBack;
+import com.messagemedia.api.http.client.OkClient;
+import com.messagemedia.api.http.response.HttpResponse;
+
+public abstract class BaseController {
+    /**
+     * Private variable to keep shared reference of client instance
+     */
+    private static HttpClient clientInstance = null;
+    private static final Object syncObject = new Object();
+    protected static final String userAgent = "APIMATIC 2.0";
+
+    /**
+     * Protected variable to keep reference of httpCallBack instance if user provides any
+     */
+    protected HttpCallBack httpCallBack = null;
+    
+    /**
+     * Get httpCallBack associated with this controller
+     * @return HttpCallBack
+     */
+    public HttpCallBack getHttpCallBack() {
+        return httpCallBack;
+    }
+    
+    /**
+     * Set the httpCallBack for this controller
+     * @param httpCallBack
+     */
+    public void setHttpCallBack(HttpCallBack httpCallBack) {
+        this.httpCallBack = httpCallBack;
+    }
+
+    /**
+     * Shared instance of the Http client
+     * @return The shared instance of the http client 
+     */
+    public static HttpClient getClientInstance() {
+        if (null == clientInstance) {
+            synchronized (syncObject) {
+                if (null == clientInstance) {
+                    clientInstance = OkClient.getSharedInstance();
+                }
+            }
+        }
+        return clientInstance;
+    }
+
+    /**
+     * Shared instance of the Http client
+     * @param    client    The shared instance of the http client 
+     */
+    public static void setClientInstance(HttpClient client) {
+        if (null != client) {
+            synchronized (syncObject) {
+                if (null != client) {
+                    clientInstance = client;
+                }
+            }
+        }
+    }
+
+    /**
+     * Validates the response against HTTP errors defined at the API level
+     * @param   response    The response recieved
+     * @param   context     Context of the request and the recieved response 
+     */
+    protected void validateResponse(HttpResponse response, HttpContext context) 
+            throws APIException {
+        //get response status code to validate
+        int responseCode = response.getStatusCode();
+        if ((responseCode < 200) || (responseCode > 208)) //[200,208] = HTTP OK
+            throw new APIException("HTTP Response Not OK", context);
+    }
+}
