+++ src/main/java/com/messagemedia/api/http/client/HttpCallBack.java
@@ -1,0 +1,20 @@
+package com.messagemedia.api.http.client;
+
+import com.messagemedia.api.http.request.HttpRequest;
+
+/**
+ * Callback to be called before and after the HTTP call for an endpoint is made 
+ */
+public interface HttpCallBack {
+    /**
+     * Callback called just before the HTTP request is sent 
+     * @param request The HTTP request to be executed
+     */
+    public void OnBeforeRequest(HttpRequest request);
+    
+    /**
+     * Callback called just after the HTTP response is received
+     * @param context Context for the HTTP call
+     */
+    public void OnAfterResponse(HttpContext context);
+}
