+++ src/main/java/com/messagemedia/api/http/client/HttpContext.java
@@ -1,0 +1,30 @@
+package com.messagemedia.api.http.client;
+
+import com.messagemedia.api.http.request.HttpRequest;
+import com.messagemedia.api.http.response.HttpResponse;
+
+public class HttpContext {
+    private HttpRequest _request;
+    private HttpResponse _response;
+
+    public HttpContext(HttpRequest request, HttpResponse response) {
+        _request = request;
+        _response = response;
+    }
+
+    /**
+     * Getter for the Http Request
+     * @return HttpRequest request
+     */
+    public HttpRequest getRequest() {
+        return _request;
+    }
+
+    /**
+     * Getter for the Http Response
+     * @return HttpResponse response
+     */
+    public HttpResponse getResponse() {
+        return _response;
+    }
+}
