+++ src/main/java/com/messagemedia/api/http/request/HttpMethod.java
@@ -1,0 +1,13 @@
+package com.messagemedia.api.http.request;
+
+/**
+ * HTTP methods enumeration
+ */
+public enum HttpMethod {
+    GET,
+    POST,
+    PUT,
+    PATCH,
+    DELETE,
+    HEAD
+}
