+++ src/main/java/com/messagemedia/api/models/Carrier.java
@@ -1,0 +1,27 @@
+package com.messagemedia.api.models;
+
+import java.util.*;
+import com.fasterxml.jackson.annotation.JsonGetter;
+import com.fasterxml.jackson.annotation.JsonSetter;
+
+public class Carrier 
+        implements java.io.Serializable {
+    private static final long serialVersionUID = -78903481635549854L;
+    private String name;
+    /** GETTER
+     * TODO: Write general description for this method
+     */
+    @JsonGetter("name")
+    public String getName ( ) { 
+        return this.name;
+    }
+    
+    /** SETTER
+     * TODO: Write general description for this method
+     */
+    @JsonSetter("name")
+    public void setName (String value) { 
+        this.name = value;
+    }
+ 
+}
