--- src/main/java/com/messagemedia/lookups/Configuration.java
@@ -1,25 +1,0 @@
-/*
- * MessageMediaLookups
- */
-package com.messagemedia.lookups;
-
-
-public class Configuration {
-    //The base Uri for API calls
-    public static String baseUri = "https://api.messagemedia.com";
-
-    //The username to use with basic authentication
-    //TODO: Replace the basicAuthUserName with an appropriate value
-    public static String basicAuthUserName = "TODO: Replace";
-
-    //The password to use with basic authentication
-    //TODO: Replace the basicAuthPassword with an appropriate value
-    public static String basicAuthPassword = "TODO: Replace";
-
-    //The username to use with HMAC authentication
-    public static String hmacAuthUserName = "";
-
-    //The password to use with HMAC authentication
-    public static String hmacAuthPassword = "";
-
-}
