--- src/main/java/com/messagemedia/lookups/controllers/BaseController.java
@@ -1,78 +1,0 @@
-/*
- * MessageMediaLookups
- */
-package com.messagemedia.lookups.controllers;
-
-import com.messagemedia.lookups.exceptions.*;
-import com.messagemedia.lookups.http.client.HttpClient;
-import com.messagemedia.lookups.http.client.HttpContext;
-import com.messagemedia.lookups.http.client.HttpCallBack;
-import com.messagemedia.lookups.http.client.UnirestClient;
-import com.messagemedia.lookups.http.response.HttpResponse;
-
-public abstract class BaseController {
-    /**
-     * Private variable to keep shared reference of client instance
-     */
-    private static HttpClient clientInstance = null;
-    private static Object syncObject = new Object();
-
-    /**
-     * Protected variable to keep reference of httpCallBack instance if user provides any
-     */
-    protected HttpCallBack httpCallBack = null;
-    
-    /**
-     * Get httpCallBack associated with this controller
-     * @return HttpCallBack
-     */
-    public HttpCallBack getHttpCallBack() {
-        return httpCallBack;
-    }
-    
-    /**
-     * Set the httpCallBack for this controller
-     * @param httpCallBack
-     */
-    public void setHttpCallBack(HttpCallBack httpCallBack) {
-        this.httpCallBack = httpCallBack;
-    }
-
-    /**
-     * Shared instance of the Http client
-     * @return The shared instance of the http client 
-     */
-    public static HttpClient getClientInstance() {
-        synchronized (syncObject) {
-            if (null == clientInstance) {
-                clientInstance = UnirestClient.getSharedInstance();
-    }
-        }
-        return clientInstance;
-    }
-
-    /**
-     * Shared instance of the Http client
-     * @param    client    The shared instance of the http client 
-     */
-    public static void setClientInstance(HttpClient client) {
-        synchronized (syncObject) {
-            if (null != client) {
-                clientInstance = client;
-            }
-        }
-    }
-
-    /**
-     * Validates the response against HTTP errors defined at the API level
-     * @param   response    The response recieved
-     * @param   context     Context of the request and the recieved response 
-     */
-    protected void validateResponse(HttpResponse response, HttpContext context) 
-            throws APIException {
-        //get response status code to validate
-        int responseCode = response.getStatusCode();
-        if ((responseCode < 200) || (responseCode > 208)) //[200,208] = HTTP OK
-            throw new APIException("HTTP Response Not OK", context);
-    }
-}
