--- src/main/java/com/messagemedia/lookups/http/client/HttpCallBack.java
@@ -1,23 +1,0 @@
-/*
- * MessageMediaLookups
-*/
-package com.messagemedia.lookups.http.client;
-
-import com.messagemedia.lookups.http.request.HttpRequest;
-
-/**
- * Callback to be called before and after the HTTP call for an endpoint is made
- */
-public interface HttpCallBack {
-    /**
-     * Callback called just before the HTTP request is sent
-     * @param request The HTTP request to be executed
-     */
-    public void OnBeforeRequest(HttpRequest request);
-
-    /**
-     * Callback called just after the HTTP response is received
-     * @param context Context for the HTTP call
-     */
-    public void OnAfterResponse(HttpContext context);
-}
