--- src/main/java/com/messagemedia/lookups/http/request/HttpMethod.java
@@ -1,17 +1,0 @@
-/*
- * MessageMediaLookups
- *
- * This file was automatically generated for MessageMedia by APIMATIC v2.0 ( https://apimatic.io ).
- */
-package com.messagemedia.lookups.http.request;
-
-/**
- * HTTP methods enumeration
- */
-public enum HttpMethod {
-    GET,
-    POST,
-    PUT,
-    PATCH,
-    DELETE
-}
